change: move Title component to shared folder
